* commit 'ada4bba910ba24402206bd72d7f14ad69c2d47b6': (268 commits)
  BLOCKS:  Added some more config items
  Fixed MinGW compile error
  Fixed a duplicate ColourId in SidePanel
  Windows:  Fix compatibility with MinGW
  VST3:  Fixed an issue in checkSizeConstraint() when the editor has a scale factor applied
  Standalone plug-in:  Fixed a mute button race condition
  VST2:  Avoided a race condition
  Replaced select() calls with poll()
  Fixed visibility for some AudioProcessorParameterWithID methods
  Linux:  Return 1.0 for GNOME scaling-factor if it hasn't been set
  Projucer:  Disabled default local code signing in Xcode
  Reverted 1ff99ec
  VST:  Fixed a race condition
  Fixed a warning on iOS
  Fixed a bug causing the Bluetooth MIDI pairing dialog to be invisible on iOS
  Re-saved all projects
  Enabled Bluetooth permissions for iOS in the DemoRunner
  Projucer:  Added Bluetooth permissions setting needed for iOS 13.0+
  VST3:  Fixed a compiler warning
  iOS:  Added an assertion if attempting to use MIDI in an iOS 13 simulator
  ...

+++ examples/GUI/BouncingBallWavetableDemo.h
@@ -217,7 +217,7 @@ public:
 
         for (auto i = 0; i < steps; ++i)
         {
-            auto p = start + ((finish - start) * i) / steps;
+            auto p = start + ((finish - start) * i) / (int) steps;
 
             auto index = (bufferIndex + i) % wavetableSize;
             waveValues[1][index] = yToAmplitude (p.y);
